raftstore: allow querying if a peer is hibernated
This PR adds a field to the response of status query /region/id that
indicates if the queried peer is hibernated.
