+++ components/raftstore/src/store/mod.rs
@@ -29,6 +29,7 @@ pub use self::bootstrap::{
 pub use self::compaction_guard::CompactionGuardGeneratorFactory;
 pub use self::config::Config;
 pub use self::fsm::{DestroyPeerJob, RaftRouter, StoreInfo};
+pub use self::hibernate_state::GroupState;
 pub use self::msg::{
     Callback, CasualMessage, ExtCallback, MergeResultKind, PeerMsg, PeerTicks, RaftCommand,
     ReadCallback, ReadResponse, SignificantMsg, StoreMsg, StoreTick, WriteCallback, WriteResponse,
